Dark mode no longer ugly
